# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "reverie",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/authorize/refresh": {
      "post": {
        "summary": "認証トークンを更新",
        "deprecated": false,
        "description": "",
        "operationId": "refreshAuthorization",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "type": "string"
                  }
                },
                "required": [
                  "refresh_token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "description": "データ",
                      "$ref": "#/components/schemas/Authorization"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/authorize/line": {
      "post": {
        "summary": "LINE でログイン",
        "deprecated": false,
        "description": "",
        "operationId": "authorizeWithLine",
        "tags": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "description": "データ",
                      "$ref": "#/components/schemas/Authorization"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/authorize/username": {
      "post": {
        "summary": "ユーザー名でログイン（テスト用）",
        "deprecated": false,
        "description": "",
        "operationId": "authorizeWithUsername",
        "tags": [],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "名前",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "description": "データ",
                      "$ref": "#/components/schemas/Authorization"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/posts/{post_id}/favorites": {
      "post": {
        "summary": "投稿にいいねする",
        "deprecated": false,
        "description": "",
        "operationId": "favoritePost",
        "tags": [],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response",
                  "required": [
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "summary": "投稿のいいねを解除する",
        "deprecated": false,
        "description": "",
        "operationId": "unfavoritePost",
        "tags": [],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response",
                  "required": [
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "summary": "投稿にいいねしたユーザーを取得する",
        "deprecated": false,
        "description": "",
        "operationId": "GetPostFavorites",
        "tags": [],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PostFavorite"
                      },
                      "description": "データ"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/posts": {
      "post": {
        "summary": "投稿を作成する",
        "deprecated": false,
        "description": "",
        "operationId": "createPost",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "reply_to_post_id": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Post"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/posts/{post_id}": {
      "get": {
        "summary": "投稿を取得する",
        "deprecated": false,
        "description": "",
        "operationId": "getPostByID",
        "tags": [],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Post"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "summary": "投稿を削除する",
        "deprecated": false,
        "description": "",
        "operationId": "deletePost",
        "tags": [],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "description": "データ"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/call_timeline": {
      "get": {
        "summary": "オープンの通話タイムラインを取得する",
        "deprecated": false,
        "description": "",
        "operationId": "getCallTimeline",
        "tags": [],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "次のページを取得するためのキー",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "description": "データ",
                      "$ref": "#/components/schemas/CallTimeline"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/call_timeline/following": {
      "get": {
        "summary": "フォロー中の通話タイムラインを取得する",
        "deprecated": false,
        "description": "",
        "operationId": "getFollowingCallTimeline",
        "tags": [],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "次のページを取得するためのキー",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "description": "データ",
                      "$ref": "#/components/schemas/CallTimeline"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/call_timeline/users/{user_id}": {
      "get": {
        "summary": "ユーザーの通話タイムラインを取得する",
        "deprecated": false,
        "description": "",
        "operationId": "getUserCallTimeline",
        "tags": [],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "次のページを取得するためのキー",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "description": "データ",
                      "$ref": "#/components/schemas/CallTimeline"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID番号",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "名前"
          },
          "nickname": {
            "type": "string"
          },
          "avatar_image_url": {
            "type": "string",
            "nullable": true
          },
          "banner_image_url": {
            "type": "string",
            "nullable": true
          },
          "biography": {
            "type": "string",
            "nullable": true
          },
          "is_private": {
            "type": "boolean"
          },
          "block_status": {
            "$ref": "#/components/schemas/BlockStatus"
          },
          "social_connection": {
            "$ref": "#/components/schemas/SocialConnection"
          },
          "social_engagement": {
            "$ref": "#/components/schemas/SocialEngagement"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "nickname",
          "avatar_image_url",
          "banner_image_url",
          "biography",
          "updated_at",
          "created_at",
          "is_private"
        ]
      },
      "SocialConnection": {
        "type": "object",
        "properties": {
          "following": {
            "type": "boolean"
          },
          "followed_by": {
            "type": "boolean"
          }
        },
        "required": [
          "following",
          "followed_by"
        ]
      },
...
      "BlockStatus": {
        "type": "object",
        "properties": {
          "blocking": {
            "type": "boolean"
          },
          "blocked_by": {
            "type": "boolean"
          }
        },
        "required": [
          "blocking",
          "blocked_by"
        ]
      },
      "SocialEngagement": {
        "type": "object",
        "properties": {
          "following_count": {
            "type": "integer"
          },
          "followers_count": {
            "type": "integer"
          },
          "posts_count": {
            "type": "integer"
          },
          "media_count": {
            "type": "integer"
          },
          "favorites_count": {
            "type": "integer"
          }
        },
        "required": [
          "following_count",
          "followers_count",
          "posts_count",
          "media_count",
          "favorites_count"
        ]
      },
      "Authorization": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "is_new": {
            "type": "boolean"
          }
        },
        "required": [
          "user_id",
          "access_token",
          "refresh_token",
          "is_new"
        ]
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID番号",
            "format": "uuid"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "favorited": {
            "type": "boolean"
          },
          "favorites_count": {
            "type": "integer"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "author",
          "updated_at",
          "created_at",
          "text",
          "favorites_count",
          "favorited"
        ]
      },
      "PostFavorite": {
        "type": "object",
        "properties": {
          "post_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "post_id",
          "created_at",
          "user"
        ]
      },
...
      "CallTimeline": {
        "type": "object",
        "properties": {
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CallRoom"
            }
          },
          "next_cursor": {
            "type": "string",
            "format": "uuid",
            "description": "次のページを取得するためのキー"
          }
        },
        "required": [
          "rooms",
          "next_cursor"
        ]
      },
...
      "CallRoom": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID番号",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "voice",
              "video"
            ]
          },
          "is_active": {
            "type": "boolean"
          },
          "joinable_by": {
            "type": "string",
            "enum": [
              "all",
              "followers",
              "friends",
              "nobody"
            ]
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CallParticipant"
            }
          }
        },
        "required": [
          "id",
          "is_active",
          "type",
          "participants",
          "title",
          "joinable_by"
        ]
      },
      "CallParticipant": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "role": {
            "type": "string",
            "enum": [
              "host",
              "co-host",
              "participant"
            ]
          }
        },
        "required": [
          "user",
          "role"
        ]
      },
      "Response": {
        "title": "",
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "正常に処理を終了したかどうか"
          },
          "code": {
            "type": "integer",
            "description": "レスポンスコード"
          },
          "data": {
            "type": "object",
            "properties": {},
            "description": "データ"
          }
        },
        "required": [
          "ok",
          "code",
          "data"
        ]
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": []
}